+++ jannovar-vardbs/src/main/java/de/charite/compbio/jannovar/vardbs/facade/DBVariantContextAnnotatorFactory.java
@@ -3,9 +3,9 @@ package de.charite.compbio.jannovar.vardbs.facade;
 import de.charite.compbio.jannovar.vardbs.base.DBAnnotationOptions;
 import de.charite.compbio.jannovar.vardbs.base.JannovarVarDBException;
 import de.charite.compbio.jannovar.vardbs.clinvar.ClinVarAnnotationDriver;
+import de.charite.compbio.jannovar.vardbs.cosmic.CosmicAnnotationDriver;
 import de.charite.compbio.jannovar.vardbs.dbsnp.DBSNPAnnotationDriver;
 import de.charite.compbio.jannovar.vardbs.exac.ExacAnnotationDriver;
-import de.charite.compbio.jannovar.vardbs.uk10k.UK10KAnnotationDriver;
 import htsjdk.variant.variantcontext.VariantContext;
 
 /**
@@ -66,7 +66,7 @@ public class DBVariantContextAnnotatorFactory {
 	 */
 	public DBVariantContextAnnotator constructUK10K(String vcfDBPath, String fastaRefPath, DBAnnotationOptions options)
 			throws JannovarVarDBException {
-		return new DBVariantContextAnnotator(new UK10KAnnotationDriver(vcfDBPath, fastaRefPath, options), options);
+		return new DBVariantContextAnnotator(new CosmicAnnotationDriver(vcfDBPath, fastaRefPath, options), options);
 	}
 
 	/**
@@ -85,6 +85,25 @@ public class DBVariantContextAnnotatorFactory {
 	public DBVariantContextAnnotator constructClinVar(String vcfDBPath, String fastaRefPath,
 			DBAnnotationOptions options) throws JannovarVarDBException {
 		return new DBVariantContextAnnotator(new ClinVarAnnotationDriver(vcfDBPath, fastaRefPath, options), options);
+	}
+
+
+	/**
+	 * Construct Cosmic {@link VariantContext} annotator factory.
+	 *
+	 * @param vcfDBPath
+	 *            Path to gzip-compressed, normalized and tbi-indexed Cosmic file to use for the annotation
+	 * @param fastaRefPath
+	 *            Path to reference FASTA file
+	 * @param options
+	 *            Configuration for the variant context annotation
+	 * @return Preconfigured {@link DBVariantContextAnnotator} object
+	 * @throws JannovarVarDBException
+	 *             on problems loading the resources
+	 */
+	public DBVariantContextAnnotator constructCosmic(String vcfDBPath, String fastaRefPath,
+			DBAnnotationOptions options) throws JannovarVarDBException {
+		return new DBVariantContextAnnotator(new CosmicAnnotationDriver(vcfDBPath, fastaRefPath, options), options);
 	}
 
 }