+++ jannovar-vardbs/src/test/java/de/charite/compbio/jannovar/vardbs/cosmic/CosmicAnnotationDriverReportAlsoOverlappingTest.java
@@ -1,4 +1,4 @@
-package de.charite.compbio.jannovar.vardbs.dbsnp;
+package de.charite.compbio.jannovar.vardbs.cosmic;
 
 import java.util.ArrayList;
 import java.util.Collections;
@@ -9,27 +9,28 @@ import org.junit.Test;
 
 import com.google.common.collect.Lists;
 
+import de.charite.compbio.jannovar.vardbs.base.DBAnnotationOptions;
 import de.charite.compbio.jannovar.vardbs.base.JannovarVarDBException;
 import htsjdk.variant.variantcontext.VariantContext;
 import htsjdk.variant.vcf.VCFHeader;
 
 /**
- * Test for annotation with dbSNP with not reporting overlapping
+ * Test for annotation with UK10K reporting overlaps besides matches
  * 
  * @author <a href="mailto:manuel.holtgrewe@bihealth.de">Manuel Holtgrewe</a>
  */
-public class DBSNPAnnotationDriverReportNoOverlappingTest extends DBSNPAnnotationDriverBaseTest {
+public class CosmicAnnotationDriverReportAlsoOverlappingTest extends CosmicAnnotationDriverBaseTest {
 
 	@Before
 	public void setUpClass() throws Exception {
 		super.setUpClass();
-		options.setReportOverlapping(false);
+		options.setReportOverlapping(true);
 		options.setReportOverlappingAsMatching(false);
 	}
 
 	@Test
 	public void testAnnotateExtendHeaderWithDefaultPrefix() throws JannovarVarDBException {
-		DBSNPAnnotationDriver driver = new DBSNPAnnotationDriver(dbSNPVCFPath, fastaPath, options);
+		CosmicAnnotationDriver driver = new CosmicAnnotationDriver(cosmicVCFPath, fastaPath, options);
 
 		VCFHeader header = vcfReader.getFileHeader();
 
@@ -49,17 +50,20 @@ public class DBSNPAnnotationDriverReportNoOverlappingTest extends DBSNPAnnotatio
 		Assert.assertEquals(6, header.getIDHeaderLines().size());
 		Assert.assertEquals(0, header.getOtherHeaderLines().size());
 
-		Assert.assertNotNull(header.getInfoHeaderLine("DBSNP_COMMON"));
-		Assert.assertNotNull(header.getInfoHeaderLine("DBSNP_CAF"));
-		Assert.assertNotNull(header.getInfoHeaderLine("DBSNP_G5"));
-		Assert.assertNotNull(header.getInfoHeaderLine("DBSNP_G5A"));
-		Assert.assertNotNull(header.getInfoHeaderLine("DBSNP_IDS"));
-		Assert.assertNotNull(header.getInfoHeaderLine("DBSNP_SAO"));
+		Assert.assertNotNull(header.getInfoHeaderLine("COSMIC_CNT"));
+		Assert.assertNotNull(header.getInfoHeaderLine("COSMIC_SNP"));
+		Assert.assertNotNull(header.getInfoHeaderLine("COSMIC_IDS"));
+		Assert.assertNotNull(header.getInfoHeaderLine("COSMIC_OVL_CNT"));
+		Assert.assertNotNull(header.getInfoHeaderLine("COSMIC_OVL_SNP"));
+		Assert.assertNotNull(header.getInfoHeaderLine("COSMIC_OVL_IDS"));
 	}
 
 	@Test
 	public void testAnnotateVariantContext() throws JannovarVarDBException {
-		DBSNPAnnotationDriver driver = new DBSNPAnnotationDriver(dbSNPVCFPath, fastaPath, options);
+		DBAnnotationOptions options = DBAnnotationOptions.createDefaults();
+		options.setReportOverlapping(true);
+		options.setReportOverlappingAsMatching(false);
+		CosmicAnnotationDriver driver = new CosmicAnnotationDriver(cosmicVCFPath, fastaPath, options);
 		VariantContext vc = vcfReader.iterator().next();
 
 		Assert.assertEquals(0, vc.getAttributes().size());
@@ -67,18 +71,20 @@ public class DBSNPAnnotationDriverReportNoOverlappingTest extends DBSNPAnnotatio
 
 		VariantContext annotated = driver.annotateVariantContext(vc);
 
-		Assert.assertEquals("rs540538026", annotated.getID());
+		Assert.assertEquals("COSM814119", annotated.getID());
 
-		Assert.assertEquals(4, annotated.getAttributes().size());
+		Assert.assertEquals(6, annotated.getAttributes().size());
 		ArrayList<String> keys = Lists.newArrayList(annotated.getAttributes().keySet());
 		Collections.sort(keys);
-		Assert.assertEquals("[CAF, COMMON, G5, IDS]", keys.toString());
+		Assert.assertEquals("[CNT, IDS, OVL_CNT, OVL_IDS, OVL_SNP, SNP]", keys.toString());
 
-		Assert.assertEquals("[0.97324, 0.02676, 0.0, 0.0]", annotated.getAttributeAsString("CAF", null));
-		Assert.assertEquals("[1, 0, 0]", annotated.getAttributeAsString("G5", null));
-		Assert.assertNull(annotated.getAttributeAsString("G5A", null));
-		Assert.assertEquals("[1, 0, 0]", annotated.getAttributeAsString("COMMON", null));
-		Assert.assertEquals("[rs540538026, ., .]", annotated.getAttributeAsString("IDS", null));
+		Assert.assertEquals("34", annotated.getAttributeAsString("CNT", null));
+		Assert.assertEquals("[COSM814119, ., .]", annotated.getAttributeAsString("IDS", null));
+		Assert.assertEquals("true", annotated.getAttributeAsString("SNP", null));
+
+		Assert.assertEquals("34", annotated.getAttributeAsString("OVL_CNT", null));
+		Assert.assertEquals("[COSM814119, ., .]", annotated.getAttributeAsString("OVL_IDS", null));
+		Assert.assertEquals("true", annotated.getAttributeAsString("OVL_SNP", null));
 	}
 
 }