+++ jannovar-vardbs/src/test/java/de/charite/compbio/jannovar/vardbs/cosmic/CosmicAnnotationDriverReportNoOverlappingTest.java
@@ -1,4 +1,4 @@
-package de.charite.compbio.jannovar.vardbs.dbsnp;
+package de.charite.compbio.jannovar.vardbs.cosmic;
 
 import java.util.ArrayList;
 import java.util.Collections;
@@ -9,16 +9,17 @@ import org.junit.Test;
 
 import com.google.common.collect.Lists;
 
+import de.charite.compbio.jannovar.vardbs.base.DBAnnotationOptions;
 import de.charite.compbio.jannovar.vardbs.base.JannovarVarDBException;
 import htsjdk.variant.variantcontext.VariantContext;
 import htsjdk.variant.vcf.VCFHeader;
 
 /**
- * Test for annotation with dbSNP with not reporting overlapping
+ * Test for annotation with COSMIC reporting no overlaps
  * 
  * @author <a href="mailto:manuel.holtgrewe@bihealth.de">Manuel Holtgrewe</a>
  */
-public class DBSNPAnnotationDriverReportNoOverlappingTest extends DBSNPAnnotationDriverBaseTest {
+public class CosmicAnnotationDriverReportNoOverlappingTest extends CosmicAnnotationDriverBaseTest {
 
 	@Before
 	public void setUpClass() throws Exception {
@@ -29,7 +30,7 @@ public class DBSNPAnnotationDriverReportNoOverlappingTest extends DBSNPAnnotatio
 
 	@Test
 	public void testAnnotateExtendHeaderWithDefaultPrefix() throws JannovarVarDBException {
-		DBSNPAnnotationDriver driver = new DBSNPAnnotationDriver(dbSNPVCFPath, fastaPath, options);
+		CosmicAnnotationDriver driver = new CosmicAnnotationDriver(cosmicVCFPath, fastaPath, options);
 
 		VCFHeader header = vcfReader.getFileHeader();
 
@@ -44,22 +45,22 @@ public class DBSNPAnnotationDriverReportNoOverlappingTest extends DBSNPAnnotatio
 
 		// Check header after extension
 		Assert.assertEquals(0, header.getFilterLines().size());
-		Assert.assertEquals(6, header.getInfoHeaderLines().size());
+		Assert.assertEquals(3, header.getInfoHeaderLines().size());
 		Assert.assertEquals(0, header.getFormatHeaderLines().size());
-		Assert.assertEquals(6, header.getIDHeaderLines().size());
+		Assert.assertEquals(3, header.getIDHeaderLines().size());
 		Assert.assertEquals(0, header.getOtherHeaderLines().size());
 
-		Assert.assertNotNull(header.getInfoHeaderLine("DBSNP_COMMON"));
-		Assert.assertNotNull(header.getInfoHeaderLine("DBSNP_CAF"));
-		Assert.assertNotNull(header.getInfoHeaderLine("DBSNP_G5"));
-		Assert.assertNotNull(header.getInfoHeaderLine("DBSNP_G5A"));
-		Assert.assertNotNull(header.getInfoHeaderLine("DBSNP_IDS"));
-		Assert.assertNotNull(header.getInfoHeaderLine("DBSNP_SAO"));
+		Assert.assertNotNull(header.getInfoHeaderLine("COSMIC_CNT"));
+		Assert.assertNotNull(header.getInfoHeaderLine("COSMIC_SNP"));
+		Assert.assertNotNull(header.getInfoHeaderLine("COSMIC_IDS"));
 	}
 
 	@Test
-	public void testAnnotateVariantContext() throws JannovarVarDBException {
-		DBSNPAnnotationDriver driver = new DBSNPAnnotationDriver(dbSNPVCFPath, fastaPath, options);
+	public void testAnnotateVariantContextNoOverlaps() throws JannovarVarDBException {
+		DBAnnotationOptions options = DBAnnotationOptions.createDefaults();
+		options.setReportOverlapping(false);
+		options.setReportOverlappingAsMatching(false);
+		CosmicAnnotationDriver driver = new CosmicAnnotationDriver(cosmicVCFPath, fastaPath, options);
 		VariantContext vc = vcfReader.iterator().next();
 
 		Assert.assertEquals(0, vc.getAttributes().size());
@@ -67,18 +68,16 @@ public class DBSNPAnnotationDriverReportNoOverlappingTest extends DBSNPAnnotatio
 
 		VariantContext annotated = driver.annotateVariantContext(vc);
 
-		Assert.assertEquals("rs540538026", annotated.getID());
+		Assert.assertEquals("COSM814119", annotated.getID());
 
-		Assert.assertEquals(4, annotated.getAttributes().size());
+		Assert.assertEquals(3, annotated.getAttributes().size());
 		ArrayList<String> keys = Lists.newArrayList(annotated.getAttributes().keySet());
 		Collections.sort(keys);
-		Assert.assertEquals("[CAF, COMMON, G5, IDS]", keys.toString());
+		Assert.assertEquals("[CNT, IDS, SNP]", keys.toString());
 
-		Assert.assertEquals("[0.97324, 0.02676, 0.0, 0.0]", annotated.getAttributeAsString("CAF", null));
-		Assert.assertEquals("[1, 0, 0]", annotated.getAttributeAsString("G5", null));
-		Assert.assertNull(annotated.getAttributeAsString("G5A", null));
-		Assert.assertEquals("[1, 0, 0]", annotated.getAttributeAsString("COMMON", null));
-		Assert.assertEquals("[rs540538026, ., .]", annotated.getAttributeAsString("IDS", null));
+		Assert.assertEquals("34", annotated.getAttributeAsString("CNT", null));
+		Assert.assertEquals("[COSM814119, ., .]", annotated.getAttributeAsString("IDS", null));
+		Assert.assertEquals("true", annotated.getAttributeAsString("SNP", null));
 	}
 
 }