+++ CHANGELOG.md
@@ -50,6 +50,8 @@
 
 ### jannovar-vardbs
 
+* Also writing out variant allele origin for dbSNP
+* Adding annotation with COSMIC
 * Fixing header description for exac database
 * Fixing output of `DBSNP_CAF` to also contain reference allele AF
 * Adding annotation with ClinVar, can annotate all clinvar variants